separated submitData from handleSubmit from Form into dataFlow
